move gene search into its own component
