+++ react/src/pages/VariantQueryPage.tsx
@@ -1,10 +1,5 @@
-import React, { useEffect, useState } from 'react';
-import { useAsyncDebounce } from 'react-table';
-import {
-    useFetchAutocompleteQuery,
-    useFetchEnsemblIdQuery,
-    useFetchVariantsQuery,
-} from '../apollo/hooks';
+import React from 'react';
+import { useFetchVariantsQuery } from '../apollo/hooks';
 import {
     Body,
     Button,
@@ -12,11 +7,13 @@ import {
     Column,
     ComboBox,
     Flex,
+    GeneSearch,
     Input,
     Spinner,
     Table,
     Typography,
 } from '../components';
+import { GeneOption } from '../components/GeneSearch';
 import { useFormReducer } from '../hooks';
 import { formIsValid, FormState, Validator } from '../hooks/useFormReducer';
 import { DropdownItem, VariantQueryResponse, VariantQueryResponseSchemaTableRow } from '../types';
@@ -198,7 +195,7 @@ const VariantQueryPage: React.FC<{}> = () => {
                                 updateQueryOptionsForm('gene')(term);
                                 updateQueryOptionsForm('ensemblId')('');
                             }}
-                            onSelect={({ ensemblId, name }: GeneModel) => {
+                            onSelect={({ ensemblId, name }: GeneOption) => {
                                 updateQueryOptionsForm('gene')(name);
                                 updateQueryOptionsForm('ensemblId')(ensemblId);
                             }}
@@ -243,77 +240,3 @@ const prepareData = (queryResponse: VariantQueryResponse) =>
     );
 
 export default VariantQueryPage;
-
-interface GeneModel {
-    ensemblId: string;
-    name: string;
-}
-
-interface GeneSearchProps {
-    value: GeneModel;
-    onSelect: (gene: GeneModel) => void;
-    onSearch: (name: string) => void;
-}
-
-const GeneSearch: React.FC<GeneSearchProps> = ({ onSearch, onSelect, value: { name } }) => {
-    const [options, setOptions] = useState<DropdownItem[]>([]);
-    const [searchTerm, setSearchTerm] = useState('');
-    const [selectedTerm, setSelectedTerm] = useState('');
-
-    const [fetchAutocompleteResults, { data: autocompleteResults, loading: autocompleteLoading }] =
-        useFetchAutocompleteQuery();
-
-    const [fetchEnsemblIdResults, { data: ensemblIdResults, loading: ensemblIdLoading }] =
-        useFetchEnsemblIdQuery();
-
-    const debouncedAutocompleteFetch = useAsyncDebounce(fetchAutocompleteResults, 500);
-
-    useEffect(() => {
-        if (!searchTerm) {
-            setOptions([]);
-        }
-        if (searchTerm.length > 2) {
-            debouncedAutocompleteFetch({ variables: { term: searchTerm.toLowerCase() } });
-        }
-    }, [debouncedAutocompleteFetch, searchTerm]);
-
-    useEffect(() => {
-        if (autocompleteResults) {
-            setOptions(
-                (autocompleteResults.autocompleteResults.suggestions || []).map(
-                    (s: { suggestion: string }, i: number) => ({
-                        value: s.suggestion.toUpperCase(),
-                        id: i,
-                        label: s.suggestion.toUpperCase(),
-                    })
-                )
-            );
-        }
-    }, [autocompleteResults]);
-
-    useEffect(() => {
-        if (ensemblIdResults && selectedTerm) {
-            onSelect({
-                ensemblId: ensemblIdResults.fetchEnsemblIdResults.entries[0]['id'],
-                name: selectedTerm,
-            });
-            setSelectedTerm('');
-        }
-    }, [ensemblIdResults, onSelect, selectedTerm]);
-
-    return (
-        <ComboBox
-            items={options}
-            loading={autocompleteLoading || ensemblIdLoading}
-            onChange={term => setSearchTerm(term)}
-            onClose={() => setOptions([])}
-            onSelect={item => {
-                setSearchTerm(item.label);
-                setSelectedTerm(item.label);
-                fetchEnsemblIdResults({ variables: { query: item.value } });
-            }}
-            placeholder="Gene Search"
-            value={name || ''}
-        />
-    );
-};
